moved fetchV2 to utils package
